🛠️Updated tweet saving and query input

+++ tweet_scraper/pipelines.py
@@ -23,28 +23,14 @@ class SaveToFilePipeline(object):
 
     def process_item(self, item, spider):
         if isinstance(item, Tweet):
-            save_path = os.path.join(self.save_tweet_path, item['id_'])
-            if os.path.isfile(save_path):
-                pass  # simply skip existing items
-                # logger.debug("skip tweet:%s"%item['id_'])
-                # or you can rewrite the file, if you don't want to skip:
-                # self.save_to_file(item,savePath)
-                # logger.debug("Update tweet:%s"%item['id_'])
-            else:
-                self.save_to_file(item, save_path)
-                logger.debug("Add tweet:%s" % item['id_'])
+            save_path = os.path.join(self.save_tweet_path, 'tweets.json')
+            self.save_to_file(item, save_path)
+            logger.debug("Add tweet:%s" % item['id_'])
 
         elif isinstance(item, User):
-            save_path = os.path.join(self.save_user_path, item['id_'])
-            if os.path.isfile(save_path):
-                pass  # simply skip existing items
-                # logger.debug("skip user:%s"%item['id_'])
-                # or you can rewrite the file, if you don't want to skip:
-                # self.save_to_file(item,savePath)
-                # logger.debug("Update user:%s"%item['id_'])
-            else:
-                self.save_to_file(item, save_path)
-                logger.debug("Add user:%s" % item['id_'])
+            save_path = os.path.join(self.save_user_path, 'users.json')
+            self.save_to_file(item, save_path)
+            logger.debug("Add user:%s" % item['id_'])
 
         else:
             logger.info("Item type is not recognized! type = %s" % type(item))
@@ -56,5 +42,13 @@ class SaveToFilePipeline(object):
                 fname - where to save
         """
         s3 = S3FileSystem(anon=False, key=SETTINGS['ACCESS_KEY_ID'], secret=SETTINGS['SECRET_ACCESS_KEY'])
-        with s3.open(fname, 'w', region='us-east-1') as file:
-            json.dump(dict(item), file, ensure_ascii=False)
+        if not s3.exists(fname):
+            with s3.open(fname, 'w', region='us-east-1') as file:
+                json.dump([{}], file)
+        with s3.open(fname, 'r', encoding='utf-8', region='us-east-1') as file:
+            org_file = json.load(file)
+            org_file.append(dict(item))
+        if not any(d.get('id_') == item['id_'] for d in org_file):
+            return None
+        with s3.open(fname, 'w', encoding='utf-8', region='us-east-1') as file:
+            json.dump(org_file, file, ensure_ascii=False)
